✨feat: Add functionality to edit reviews

+++ app/Http/Controllers/ReviewController.php
@@ -50,20 +50,21 @@ class ReviewController extends Controller
     // 1. 데이터베이스에서 해당 리뷰 조회
     // 전달 받은 id값과 일치하는 레코드를 $review에 저장한다.
     $review = Review::find($id);
-
     // 리뷰가 존재하지 않는 경우 오류 응답 반환
     // find() 메서드는 값이 존재하지 않으면 null 값을 반환하기에 만약 해당하는 레코드가 없으면 에러를 전달
     if (!$review) {
         return response()->json(['message' => 'Review not found'], 404);
     }
-
     // 2. 리뷰 삭제 실행
-    $review->delete();
+    if(Auth::id()==$review['author_id']) {
+        $review->delete();
+    }
     // 데이터베이스에 해당 값이 삭제되었는지 확인
     // 3. 성공 응답 반환
     return response()->json(['message' => 'Review deleted successfully']);
     }
-
+    
+    // 리뷰 수정을 위해 아이디 값으로 리뷰 반환
     public function getReviewById($id) 
     {
         // Review:: 는 Review 모델에 대한 정적인 호출을 의미
@@ -81,4 +82,31 @@ class ReviewController extends Controller
         // 일치하면 반환
         return response()->json(['review' => $review]);
     }
+
+    // 리뷰 수정
+    public function editReview(Request $request)
+    {        
+        // request 변수로 입력 정보를 받아서 validate에 정의 된 규칙에 부합한지 판단하고
+        // $validate 변수에 저장
+        $validated = $request->validate([
+            'id'=>'required',
+            'rating'=>'required',
+            'review_text'=>'required|min:10',
+            'image_file'=>'required',
+            ]);
+        try {
+            $review = Review::find($validated['id']);
+            // 정보 갱신
+            $review->rating = $validated['rating'];
+            $review->review_text = $validated['review_text'];
+            $file = $request->file('image_file'); // $file에 저장
+            $path = $file->store('photos'); // store()메서드는 지정된 경로에 파일을 저장
+            $review->image_file = $path; // store 메서드를 활용해 저장한 경로가 $path 변수에 할당됨
+            $review->save();
+            return response()->json(['message' => 'Edit review successfully']);
+        } catch (\Exception $e) {
+            // 에러 발생 시
+            return response()->json(['message' => 'Failed to edit review'], 500);
+        }
+    }    
 }
